fixed deprecated np.float and np.complex

+++ petram/phys/em1d/em1d_model.py
@@ -274,8 +274,8 @@ class EM1D(PhysModule):
         from petram.helper.variables import TestVariable
 
         freq, omega = self.get_freq_omega()
-        add_constant(v, 'omega', suffix, np.float(omega),)
-        add_constant(v, 'freq', suffix, np.float(freq),)
+        add_constant(v, 'omega', suffix, np.float64(omega),)
+        add_constant(v, 'freq', suffix, np.float64(freq),)
         add_constant(v, 'mu0', '', self._global_ns['mu0'])
         add_constant(v, 'e0', '', self._global_ns['e0'])
